packages/YT_test/requirements.txt
youtube-dl
google-cloud-core
google-cloud-storage
gswrap
pyyaml
